refactor(#163): api 파일 폴더 하위로 구조 변경
