+++ src/api/reissue/reissue.ts
@@ -1,6 +1,6 @@
 import { getRefreshToken } from "@/utils/storage";
-import Axios from ".";
 import { ReissueResponse } from "@/types/api/reissue/reissue";
+import Axios from "..";
 
 export const reissueToken = async ():Promise<ReissueResponse> => {
   const endpoint = "/api/v2/auth/refresh";
