--- src/interface/report.ts
@@ -1,5 +1,0 @@
-export interface Report {
-    targetMemberId: number;
-    reportTypeIdList: number[];
-    contents: string;
-}
